==> MyWork/Tianchi/Fianl_Follow/Step1.1_preprocess_luna16.ipynb <==
from utils.imports import *

def find_mhd_file(data_path,patient_id):
    for subject_no in range(LUNA_SUBSET_START_INDEX, 15):
        src_dir = data_path + "*_subset" + str(subject_no).zfill(2) + "/"
        for src_path in glob.glob(src_dir + "*.mhd"):
            if patient_id in src_path:
                return src_path
    return None


def normalize(image):
    MIN_BOUND = -1000.0
    MAX_BOUND = 400.0
    image = (image - MIN_BOUND) / (MAX_BOUND - MIN_BOUND)
    image[image > 1] = 1.
    image[image < 0] = 0.
    return image


def process_image(src_path):
    patient_id = ntpath.basename(src_path).replace(".mhd", "")
    dst_dir = pic_path + patient_id + "/"
    itk_img = SimpleITK.ReadImage(src_path)
    img_array = SimpleITK.GetArrayFromImage(itk_img)
    origin = numpy.array(itk_img.GetOrigin())      # x,y,z  Origin in world coordinates (mm)
    direction = numpy.array(itk_img.GetDirection())      # x,y,z  Origin in world coordinates (mm)
    spacing = numpy.array(itk_img.GetSpacing())    # spacing of voxels in world coor. (mm)
    rescale = spacing / TARGET_VOXEL_MM
    img_array = helpers.rescale_patient_images(img_array, spacing, TARGET_VOXEL_MM)
    img_list = []
    for i in range(img_array.shape[0]):
        img = img_array[i]
        seg_img, mask = helpers.get_segmented_lungs(img.copy())
        img_list.append(seg_img)
        img = normalize(img)
        cv2.imwrite(dst_dir + "img_" + str(i).rjust(4, '0') + "_i.png", img * 255)
        cv2.imwrite(dst_dir + "img_" + str(i).rjust(4, '0') + "_m.png", mask * 255)


def process_pos_annotations_patient(src_path, label_path,csv_path, patient_id):
    df_node = pandas.read_csv(csv_path + "annotations.csv")
    dst_dir = dst_dir + patient_id + "/"
    itk_img = SimpleITK.ReadImage(src_path)
    img_array = SimpleITK.GetArrayFromImage(itk_img)
    df_patient = df_node[df_node["seriesuid"] == patient_id]
    num_z, height, width = img_array.shape        #heightXwidth constitute the transverse plane
    origin = numpy.array(itk_img.GetOrigin())      # x,y,z  Origin in world coordinates (mm)
    spacing = numpy.array(itk_img.GetSpacing())    # spacing of voxels in world coor. (mm)
    rescale = spacing / TARGET_VOXEL_MM

    direction = numpy.array(itk_img.GetDirection())      # x,y,z  Origin in world coordinates (mm)
    flip_direction_x = False
    flip_direction_y = False
    if round(direction[0]) == -1:
        origin[0] *= -1
        direction[0] = 1
        flip_direction_x = True
    if round(direction[4]) == -1:
        origin[1] *= -1
        direction[4] = 1
        flip_direction_y = True
    assert abs(sum(direction) - 3) < 0.01
    patient_imgs = helpers.load_patient_images(patient_id, pic_path, "*_i.png")
    pos_annos = []
    df_patient = df_node[df_node["seriesuid"] == patient_id]
    anno_index = 0
    for index, annotation in df_patient.iterrows():
        node_x = annotation["coordX"]
        if flip_direction_x:
            node_x *= -1
        node_y = annotation["coordY"]
        if flip_direction_y:
            node_y *= -1
        node_z = annotation["coordZ"]
        diam_mm = annotation["diameter_mm"]
        center_float = numpy.array([node_x, node_y, node_z])
        center_int = numpy.rint((center_float-origin) / spacing)
        center_float_rescaled = (center_float - origin) / TARGET_VOXEL_MM
        center_float_percent = center_float_rescaled / patient_imgs.swapaxes(0, 2).shape
        diameter_pixels = diam_mm / TARGET_VOXEL_MM
        diameter_percent = diameter_pixels / float(patient_imgs.shape[1])

        pos_annos.append([anno_index, round(center_float_percent[0], 4), round(center_float_percent[1], 4), round(center_float_percent[2], 4), round(diameter_percent, 4), 1])
        anno_index += 1

    df_annos = pandas.DataFrame(pos_annos, columns=["anno_index", "coord_x", "coord_y", "coord_z", "diameter", "malscore"])
    df_annos.to_csv(label_path + patient_id + "_annos_pos.csv", index=False)
    return [patient_id, spacing[0], spacing[1], spacing[2]]



def process_images(delete_existing=False, only_process_patient=None):
    if delete_existing and os.path.exists(pic_path):
        print("Removing old stuff..")
        if os.path.exists(pic_path):
            shutil.rmtree(pic_path)

    for subject_no in range(LUNA_SUBSET_START_INDEX, 15):
        src_dir = data_path + "*_subset" + str(subject_no).zfill(2) + "/"
        src_paths = glob.glob(src_dir + "*.mhd")

        if only_process_patient is None and True:
            pool = multiprocessing.Pool(6)
            pool.map(process_image, src_paths)
        else:
            for src_path in src_paths:
                print(src_path)
                if only_process_patient is not None:
                    if only_process_patient not in src_path:
                        continue
                process_image(src_path)


def process_pos_annotations_patient2():
    candidate_index = 0
    only_patient = None
    for subject_no in range(LUNA_SUBSET_START_INDEX, 15):
        src_dir = data_path  + "*_subset" + str(subject_no).zfill(2) +  "/"
        print(src_dir)
        for src_path in glob.glob(src_dir + "*.mhd"):
            if only_patient is not None and only_patient not in src_path:
                continue
            patient_id = ntpath.basename(src_path).replace(".mhd", "")
            print(candidate_index, " patient: ", patient_id)
            process_pos_annotations_patient(src_path, label_path,csv_path, patient_id)
            candidate_index += 1

data_path = PATH['src_train']
pic_path = PATH['pic_train']
label_path = PATH['label_train']
csv_path = PATH['annotations_train']

#### LKDS-00395，LKDS-00731，LKDS-00983的尺寸为768*768，会报错
#### LKDS-00434 有错，空文件夹，LKDS-00984没文件夹
#### subset06结束的时候会卡住

LUNA_SUBSET_START_INDEX = 0

only_process_patient = None
process_images(delete_existing=False, only_process_patient=only_process_patient)

LUNA_SUBSET_START_INDEX = 14

process_pos_annotations_patient2()

## 修改计划
### 1、读取一遍文件自动完成2个处理
### 2、异常文件的调试
### 3、错误自动跳过

data_path = PATH['src_test']
pic_path = PATH['pic_test']
label_path = PATH['label_test']
csv_path = PATH['annotations_test']

#### LKDS-00186

only_process_patient = None
process_images(delete_existing=False, only_process_patient=only_process_patient)

process_pos_annotations_patient2()
